requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
revolt==1.9.0